// src/api/openapi/models/CurrentUserDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UserRole } from './UserRole';
import {
    UserRoleFromJSON,
    UserRoleFromJSONTyped,
    UserRoleToJSON,
    UserRoleToJSONTyped,
} from './UserRole';

/**
 * 
 * @export
 * @interface CurrentUserDto
 */
export interface CurrentUserDto {
    /**
     * 
     * @type {number}
     * @memberof CurrentUserDto
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof CurrentUserDto
     */
    email: string;
    /**
     * 
     * @type {UserRole}
     * @memberof CurrentUserDto
     */
    role: UserRole;
}



/**
 * Check if a given object implements the CurrentUserDto interface.
 */
export function instanceOfCurrentUserDto(value: object): value is CurrentUserDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('email' in value) || value['email'] === undefined) return false;
    if (!('role' in value) || value['role'] === undefined) return false;
    return true;
}

export function CurrentUserDtoFromJSON(json: any): CurrentUserDto {
    return CurrentUserDtoFromJSONTyped(json, false);
}

export function CurrentUserDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CurrentUserDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'email': json['email'],
        'role': UserRoleFromJSON(json['role']),
    };
}

export function CurrentUserDtoToJSON(json: any): CurrentUserDto {
    return CurrentUserDtoToJSONTyped(json, false);
}

export function CurrentUserDtoToJSONTyped(value?: CurrentUserDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'email': value['email'],
        'role': UserRoleToJSON(value['role']),
    };
}