// src/api/openapi/models/TransferInfoResponseDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TransferInfoResponseDto
 */
export interface TransferInfoResponseDto {
    /**
     * 
     * @type {string}
     * @memberof TransferInfoResponseDto
     */
    iban: string;
    /**
     * 
     * @type {string}
     * @memberof TransferInfoResponseDto
     */
    swift: string;
}

/**
 * Check if a given object implements the TransferInfoResponseDto interface.
 */
export function instanceOfTransferInfoResponseDto(value: object): value is TransferInfoResponseDto {
    if (!('iban' in value) || value['iban'] === undefined) return false;
    if (!('swift' in value) || value['swift'] === undefined) return false;
    return true;
}

export function TransferInfoResponseDtoFromJSON(json: any): TransferInfoResponseDto {
    return TransferInfoResponseDtoFromJSONTyped(json, false);
}

export function TransferInfoResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TransferInfoResponseDto {
    if (json == null) {
        return json;
    }
    return {
        
        'iban': json['iban'],
        'swift': json['swift'],
    };
}

export function TransferInfoResponseDtoToJSON(json: any): TransferInfoResponseDto {
    return TransferInfoResponseDtoToJSONTyped(json, false);
}

export function TransferInfoResponseDtoToJSONTyped(value?: TransferInfoResponseDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'iban': value['iban'],
        'swift': value['swift'],
    };
}